% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\name{adonis.pair}
\alias{adonis.pair}
\title{Pairwise comparisons for Permutational Multivariate Analysis of Variance Using Distance Matrices}
\usage{
adonis.pair(dist.mat, Factor, nper = 1000, corr.method = "fdr")
}
\arguments{
\item{dist.mat}{Dissimilarity object}

\item{Factor}{Factor whose levels are to be compared.}

\item{nper}{Number of permutations.}

\item{corr.method}{P-value's correction method (from \code{p.adjust}).}
}
\value{
Data frame with the p-values and corrected p-values for each pairwise combination.
}
\description{
Pairwise comparisons for all pairs of levels of a factor by using for Permutational MANOVA.
}
\details{
Basically the \code{adonis.pair} function applies the \code{adonis} function from \pkg{vegan} to all pairs of levels of a factor. P-values are then corrected with \code{p.adjust}.
}
\examples{
library(vegan)
data(dune)
data(dune.env)
adonis.pair(vegdist(dune),dune.env$Management)
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{EcolUtils}

