% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\docType{package}
\name{EcolUtils}
\alias{EcolUtils}
\alias{EcolUtils-package}
\title{EcolUtils: Utilities for community ecology analysis.}
\description{
The package \pkg{EcolUtils} provides tools for community ecology analysis not present in up-to-date available packages. Designed with molecular 16S-derived data in mind.
}
\details{
The package \pkg{EcolUtils} depends on \pkg{vegan} which can be installed from CRAN.

To see the preferable citation of the package, type \code{citation("EcolUtils")}.
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}

