% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\name{niche.range}
\alias{niche.range}
\title{Niche range computation for OTUs in a community through matrix randomization.}
\usage{
niche.range(comm.tab, env.var, n = 1000, probs = c(0.025, 0.975))
}
\arguments{
\item{comm.tab}{Community data, a matrix-like object (samples as rows; OTUs as columns).}

\item{env.var}{Environmental variable as a numeric vector.}

\item{n}{Number of permutations.}

\item{probs}{Probabilities for confidence interval calculations.}
}
\value{
Data frame with the observed niche range value, the mean and CI null values and the classification of each OTU based on randomizations.
}
\description{
Computation of the range of an environmental variable for all OTUs in a community and statistical comparison to randomized communities.
}
\details{
The \code{niche.val} function computes the range of values of an environmental variable where each OTU is present. The mean value and CI for each OTU is computed for \code{n} null matrices. The null matrices are created by randomly shuffling the rows in \code{comm.tab}.  Each OTU is classified depending whether the real niche range is lower / higher / within the CI.
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{EcolUtils}
