% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\name{rrarefy.perm}
\alias{rrarefy.perm}
\title{Rarefaction of a community matrix with permutations}
\usage{
rrarefy.perm(x, sample, n = 100, round.out = T)
}
\arguments{
\item{x}{Community data, a matrix-like object.}

\item{sample}{Subsample size.}

\item{n}{Number of independent rarefactions.}

\item{round.out}{logical; should output be rounded.}
}
\value{
Rarefied community.
}
\description{
This function generates one randomly rarefied community data frame through \code{n} repeated independent rarefactions.
}
\details{
Function \code{rrarefy.perm} generates one randomly rarefied community data frame by computing \code{n} rarefied communities with \code{rrarefy} in \pkg{vegan} and computing the mean.
The average value for each cell may (or may not) be rounded by using \code{round.out} parameter.
}
\examples{
data(varespec)
rrarefy.perm(varespec*100)
}
\author{
Guillem Salazar <salazar@icm.csic.es>
}
\keyword{EcolUtils}

