% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\name{smwda}
\alias{smwda}
\title{Split moving-window distance analysis}
\usage{
smwda(comm.dist.mat, env.var, w.size = 10, nrep = 1000, probs = c(0.025,
  0.975))
}
\arguments{
\item{comm.dist.mat}{Dissimilarity matrix.}

\item{env.var}{Vector representing a continuous environmental variable.}

\item{w.size}{Windows size (has to be even).}

\item{nrep}{Number of randomizations for significance computation.}

\item{probs}{Probabilities for confidence interval calculations.}
}
\value{
List containing: 1) data frame with the mean \code{env.var} value of the two central samples of the window, the min and max \code{env.var} values of the window, the statistic, its z-score and significance. 2) The mean and two quantiles of the null statistic based on the \code{probs}, all the null values and a window-to-sample map.
}
\description{
Split moving-window analysis based on multivariate community data and permutations.
}
\details{
For each window the data is divided in two halves and the mean distance between samples belonging to a different half is divided by the mean distance between samples belonging to the same half. This is used as an statistic for which \code{nrep} null statistics are computed by resampling the order of the distance matrix and significance is computed. A z-score of the statistic is also provided.
}
\examples{
library(vegan)
data("varespec")
data("varechem")
tmp<-smwda(vegdist(varespec),varechem$N)
plot(tmp$windows$env.var.mean,tmp$windows$stat.real.zscore,col=tmp$windows$sign,type="b",pch=19)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{EcolUtils}
