% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcolUtils_functions.R
\name{spec.gen.optimized}
\alias{spec.gen.optimized}
\title{Optimized Specialist/Generalist Classification of OTUs}
\usage{
spec.gen.optimized(comm.tab, niche.width.method = "levins", perm.method = "quasiswap",
  n = 1000, probs = c(0.025, 0.975))
}
\arguments{
\item{comm.tab}{Community data, a matrix-like object with samples as rows and OTUs as columns.}

\item{niche.width.method}{Niche width index: \code{levins} (default), \code{shannon}, or \code{occurrence} (the count of samples an OTU appears in).}

\item{perm.method}{Method for null model construction: \code{quasiswap} (default), from \pkg{vegan}. Optimized for efficiency.}

\item{n}{Number of permutations, with parallel processing to improve performance.}

\item{probs}{Probabilities for confidence interval calculations, typically for 95% confidence intervals.}
}
\value{
A data frame with each OTU's observed niche width, mean and CI of null values, and their classification.
}
\description{
Efficiently classifies OTUs as generalists, specialists, or non-significant by comparing niche width indexes against null values derived from permutation algorithms. This version utilizes parallel processing for enhanced performance.
}
\details{
The function calculates a niche width index for each OTU, generating a distribution of null values through permutation of the community matrix. OTUs are classified based on their real niche width's deviation from this null distribution's confidence intervals.
}
\
